create hooks and refectore the code
